+++ x/gamm/pool-models/stableswap/pool.go
@@ -11,12 +11,12 @@ import (
 	"github.com/osmosis-labs/osmosis/osmomath"
 	"github.com/osmosis-labs/osmosis/v13/x/gamm/pool-models/internal/cfmm_common"
 	"github.com/osmosis-labs/osmosis/v13/x/gamm/types"
-	swaproutertypes "github.com/osmosis-labs/osmosis/v13/x/swaprouter/types"
+	poolmanagertypes "github.com/osmosis-labs/osmosis/v13/x/poolmanager/types"
 )
 
 var (
-	_ swaproutertypes.PoolI = &Pool{}
-	_ types.CFMMPoolI       = &Pool{}
+	_ poolmanagertypes.PoolI = &Pool{}
+	_ types.CFMMPoolI        = &Pool{}
 )
 
 // NewStableswapPool returns a stableswap pool