+++ x/pool-incentives/keeper/keeper.go
@@ -22,14 +22,14 @@ type Keeper struct {
 
 	paramSpace paramtypes.Subspace
 
-	accountKeeper    types.AccountKeeper
-	bankKeeper       types.BankKeeper
-	incentivesKeeper types.IncentivesKeeper
-	distrKeeper      types.DistrKeeper
-	swaprouterKeeper types.SwapRouterKeeper
+	accountKeeper     types.AccountKeeper
+	bankKeeper        types.BankKeeper
+	incentivesKeeper  types.IncentivesKeeper
+	distrKeeper       types.DistrKeeper
+	poolmanagerKeeper types.PoolManagerKeeper
 }
 
-func NewKeeper(storeKey sdk.StoreKey, paramSpace paramtypes.Subspace, accountKeeper types.AccountKeeper, bankKeeper types.BankKeeper, incentivesKeeper types.IncentivesKeeper, distrKeeper types.DistrKeeper, swaprouterKeeper types.SwapRouterKeeper) Keeper {
+func NewKeeper(storeKey sdk.StoreKey, paramSpace paramtypes.Subspace, accountKeeper types.AccountKeeper, bankKeeper types.BankKeeper, incentivesKeeper types.IncentivesKeeper, distrKeeper types.DistrKeeper, poolmanagerKeeper types.PoolManagerKeeper) Keeper {
 	// ensure pool-incentives module account is set
 	if addr := accountKeeper.GetModuleAddress(types.ModuleName); addr == nil {
 		panic(fmt.Sprintf("%s module account has not been set", types.ModuleName))
@@ -45,11 +45,11 @@ func NewKeeper(storeKey sdk.StoreKey, paramSpace paramtypes.Subspace, accountKee
 
 		paramSpace: paramSpace,
 
-		accountKeeper:    accountKeeper,
-		bankKeeper:       bankKeeper,
-		incentivesKeeper: incentivesKeeper,
-		distrKeeper:      distrKeeper,
-		swaprouterKeeper: swaprouterKeeper,
+		accountKeeper:     accountKeeper,
+		bankKeeper:        bankKeeper,
+		incentivesKeeper:  incentivesKeeper,
+		distrKeeper:       distrKeeper,
+		poolmanagerKeeper: poolmanagerKeeper,
 	}
 }